use blocking/nonblocking reads instead of poll(2)
Previously, we used a blocking fd for the token pipe and used poll(2) to
see if anything could be read right now in try_procure().  However,
there is a race between poll(2) and the subsequent blocking read(2),
which can result in processes deadlocking when multiple instances
of redo poll at the same time.

Instead, we now make the fd nonblocking when needed, so that the read
either reads or returns immediately if there is nothing to read.

Closes #3.