Detecting infinite dependemcy loops attempts
Achieved with the help of storing the whole dependency path in REDO_TRACK
environment variable and searching the next target full path inside REDO_TRACK.
If found means an attempt for infinite loop, required branch is ommitted.
If not found, target full path is appended to REDO_TRACK and .do file is
envoked with the updated REDO_TRACK value.